+++ repository/IngSoft2-Model/back/src/main/java/api/back/TransaccionesRepository.java
@@ -2,6 +2,6 @@ package api.back;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransaccionesRepository extends JpaRepository<Transacciones, Integer> {
+public interface TransaccionesRepository extends JpaRepository<Transacciones, Long> {
 
 }
